+++ CollapseDisplay/HealthBarHooks.cs
@@ -1,8 +1,10 @@
-﻿using Mono.Cecil;
+﻿using CollapseDisplay.Config;
+using Mono.Cecil;
 using Mono.Cecil.Cil;
 using MonoMod.Cil;
 using RoR2;
 using RoR2.UI;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace CollapseDisplay
@@ -47,11 +49,14 @@ namespace CollapseDisplay
 
             c.Index = 0;
 
-            VariableDefinition collapseDisplayBarInfo = new VariableDefinition(il.Import(typeof(HealthBar.BarInfo)));
-            il.Method.Body.Variables.Add(collapseDisplayBarInfo);
+            VariableDefinition damageDisplayBarInfos = new VariableDefinition(il.Import(typeof(HealthBar.BarInfo[])));
+            il.Method.Body.Variables.Add(damageDisplayBarInfos);
+
+            const int NUM_BAR_INFOS = 2;
 
             c.Emit(OpCodes.Ldarg_0);
-            c.EmitDelegate((HealthBar healthBar) =>
+            c.EmitDelegate(getDelayDamageBarInfos);
+            static HealthBar.BarInfo[] getDelayDamageBarInfos(HealthBar healthBar)
             {
                 HealthBarType healthBarType = HealthBarType.Unknown;
                 if (healthBar.TryGetComponent(out HealthBarTypeProvider healthBarTypeProvider))
@@ -59,64 +64,125 @@ namespace CollapseDisplay
                     healthBarType = healthBarTypeProvider.Type;
                 }
 
-                HealthBarCollapseDisplayOptions displayOptions = CollapseDisplayPlugin.GetDisplayOptions(healthBarType);
-                HealthBarStyle.BarStyle style = displayOptions.IndicatorStyle;
+                int barInfoIndex = 0;
+                HealthBar.BarInfo[] barInfos = new HealthBar.BarInfo[NUM_BAR_INFOS];
 
-                HealthBar.BarInfo collapseBarInfo = new HealthBar.BarInfo
+                ref HealthBar.BarInfo requestBarInfo()
+                {
+                    return ref barInfos[barInfoIndex++];
+                }
+
+                DelayedDamageDisplayOptions collapseDisplayOptions = CollapseDisplayPlugin.CollapseDisplayOptions;
+                DelayedDamageBarStyle collapseDamageBarStyle = collapseDisplayOptions.GetDamageBarStyle(healthBarType);
+
+                HealthBarStyle.BarStyle collapseIndicatorStyle = collapseDamageBarStyle.BarStyle;
+                ref HealthBar.BarInfo collapseBarInfo = ref requestBarInfo();
+                collapseBarInfo = new HealthBar.BarInfo
                 {
                     enabled = false,
-                    color = style.baseColor,
-                    sprite = style.sprite,
-                    imageType = style.imageType,
-                    sizeDelta = style.sizeDelta,
+                    color = collapseIndicatorStyle.baseColor,
+                    sprite = collapseIndicatorStyle.sprite,
+                    imageType = collapseIndicatorStyle.imageType,
+                    sizeDelta = collapseIndicatorStyle.sizeDelta,
+                    normalizedXMin = 0f,
+                    normalizedXMax = 0
+                };
+
+                DelayedDamageDisplayOptions warpedEchoDisplayOptions = CollapseDisplayPlugin.WarpedEchoDisplayOptions;
+                DelayedDamageBarStyle warpedEchoDamageBarStyle = warpedEchoDisplayOptions.GetDamageBarStyle(healthBarType);
+
+                HealthBarStyle.BarStyle warpedEchoIndicatorStyle = warpedEchoDamageBarStyle.BarStyle;
+                ref HealthBar.BarInfo warpedEchoBarInfo = ref requestBarInfo();
+                warpedEchoBarInfo = new HealthBar.BarInfo
+                {
+                    enabled = false,
+                    color = warpedEchoIndicatorStyle.baseColor,
+                    sprite = warpedEchoIndicatorStyle.sprite,
+                    imageType = warpedEchoIndicatorStyle.imageType,
+                    sizeDelta = warpedEchoIndicatorStyle.sizeDelta,
                     normalizedXMin = 0f,
                     normalizedXMax = 0f
                 };
 
-                if (displayOptions.EnabledConfig.Value)
+                if (barInfoIndex != NUM_BAR_INFOS)
                 {
-                    HealthComponent healthComponent = healthBar.source;
-                    if (healthComponent && healthComponent.TryGetComponent(out CollapseDamageProvider collapseDamageProvider))
+                    Log.Error($"Missing registered bars, expected {NUM_BAR_INFOS}, registered {barInfoIndex}");
+                }
+
+                HealthComponent healthComponent = healthBar.source;
+                if (healthComponent && healthComponent.TryGetComponent(out DelayedDamageProvider delayedDamageProvider))
+                {
+                    HealthComponent.HealthBarValues healthBarValues = healthComponent.GetHealthBarValues();
+
+                    float currentHealth = healthComponent.combinedHealth;
+                    float fullCombinedHealth = healthComponent.fullCombinedHealth;
+
+                    float nonCurseFraction = 1f - healthBarValues.curseFraction;
+
+                    float totalBarDamageAmount = 0f;
+
+                    void addBar(ref HealthBar.BarInfo barInfo, float damageAmount)
                     {
-                        float collapseDamage = collapseDamageProvider.EstimatedCollapseDamage;
-                        if (collapseDamage > 0f)
+                        barInfo.enabled = true;
+
+                        float xMin = (currentHealth - totalBarDamageAmount - damageAmount) / fullCombinedHealth * nonCurseFraction;
+                        barInfo.normalizedXMin = Mathf.Clamp01(xMin);
+
+                        float xMax = (currentHealth - totalBarDamageAmount) / fullCombinedHealth * nonCurseFraction;
+                        barInfo.normalizedXMax = Mathf.Clamp01(xMax);
+
+                        totalBarDamageAmount += damageAmount;
+                    }
+
+                    if (warpedEchoDamageBarStyle.EnabledConfig.Value)
+                    {
+                        DelayedDamageInfo warpedEchoDamageInfo = delayedDamageProvider.WarpedEchoDamage;
+                        if (warpedEchoDamageInfo.DamageTimestamp.timeUntil > 0f)
                         {
-                            collapseBarInfo.enabled = true;
+                            float warpedEchoDisplayedDamage = warpedEchoDamageInfo.Damage - healthComponent.barrier;
+                            if (warpedEchoDisplayedDamage > 0f)
+                            {
+                                addBar(ref warpedEchoBarInfo, warpedEchoDisplayedDamage);
+                            }
+                        }
+                    }
 
-                            float fullCombinedHealth = healthComponent.fullCombinedHealth;
-                            float combinedHealth = healthComponent.health + healthComponent.shield;
-
-                            HealthBar.BarInfo instantHealthBarInfo = healthBar.barInfoCollection.instantHealthbarInfo;
-                            HealthBar.BarInfo curseBarInfo = healthBar.barInfoCollection.curseBarInfo;
-
-                            float healthBarMin = instantHealthBarInfo.normalizedXMin;
-                            float healthBarMax = curseBarInfo.enabled ? curseBarInfo.normalizedXMin : 1f;
-
-                            collapseBarInfo.normalizedXMin = Util.Remap(Mathf.Clamp01((combinedHealth - collapseDamage) / fullCombinedHealth), 0f, 1f, healthBarMin, healthBarMax);
-
-                            collapseBarInfo.normalizedXMax = Util.Remap(Mathf.Clamp01(combinedHealth / fullCombinedHealth), 0f, 1f, healthBarMin, healthBarMax);
+                    if (collapseDamageBarStyle.EnabledConfig.Value)
+                    {
+                        DelayedDamageInfo collapseDamageInfo = delayedDamageProvider.CollapseDamage;
+                        if (collapseDamageInfo.DamageTimestamp.timeUntil > 0f)
+                        {
+                            float collapseDisplayedDamage = collapseDamageInfo.Damage - healthComponent.barrier;
+                            if (collapseDisplayedDamage > 0f)
+                            {
+                                addBar(ref collapseBarInfo, collapseDisplayedDamage);
+                            }
                         }
                     }
                 }
 
-                return collapseBarInfo;
-            });
+                return barInfos;
+            }
 
-            c.Emit(OpCodes.Stloc, collapseDisplayBarInfo);
+            c.Emit(OpCodes.Stloc, damageDisplayBarInfos);
 
             if (c.TryGotoNext(MoveType.After,
                               x => x.MatchCallOrCallvirt<HealthBar.BarInfoCollection>(nameof(HealthBar.BarInfoCollection.GetActiveCount))))
             {
-                c.Emit(OpCodes.Ldloca, collapseDisplayBarInfo);
-                c.EmitDelegate((int activeCount, in HealthBar.BarInfo collapseDisplayBarInfo) =>
+                c.Emit(OpCodes.Ldloc, damageDisplayBarInfos);
+                c.EmitDelegate(addDamageDisplayBarsToCount);
+                static int addDamageDisplayBarsToCount(int activeCount, HealthBar.BarInfo[] damageDisplayBarInfos)
                 {
-                    if (collapseDisplayBarInfo.enabled)
+                    for (int i = 0; i < damageDisplayBarInfos.Length; i++)
                     {
-                        activeCount++;
+                        if (damageDisplayBarInfos[i].enabled)
+                        {
+                            activeCount++;
+                        }
                     }
 
                     return activeCount;
-                });
+                }
             }
             else
             {
@@ -131,13 +197,18 @@ namespace CollapseDisplay
                 return;
             }
 
-            c.Emit(OpCodes.Ldarg_0);
+            for (int i = 0; i < NUM_BAR_INFOS; i++)
+            {
+                c.Emit(OpCodes.Ldarg_0);
 
-            c.Emit(OpCodes.Ldloca, collapseDisplayBarInfo);
+                c.Emit(OpCodes.Ldloc, damageDisplayBarInfos);
+                c.Emit(OpCodes.Ldc_I4, i);
+                c.Emit(OpCodes.Ldelema, il.Import(typeof(HealthBar.BarInfo)));
 
-            c.Emit(OpCodes.Ldloca, localsVar);
+                c.Emit(OpCodes.Ldloca, localsVar);
 
-            c.Emit(OpCodes.Call, handleBarMethod);
+                c.Emit(OpCodes.Call, handleBarMethod);
+            }
         }
     }
 }